\name{DrupalR-package}
\alias{DrupalR-package}
\alias{DrupalR}
\docType{package}
\title{
Protypes for accessing Drupal websites.
}
\description{
Authenticate to and get content from Drupal websites.
}
\details{
\tabular{ll}{
Package: \tab DrupalR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-13\cr
License: \tab GPL\cr
}
Helper functionality for other packages, such as BioAcousticaR.
}
\author{
Ed Baker

Maintainer: Ed Baker <edwbaker@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
