\name{DrupalR-package}
\alias{DrupalR-package}
\alias{DrupalR}
\docType{package}
\title{
Protypes for accessing Drupal websites.
}
\description{
Authenticate to and get content from Drupal websites from within the R environment.
}
\details{
\tabular{ll}{
Package: \tab DrupalR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-13\cr
License: \tab GPL\cr
}
This module provides helper functionality for other packages, such as BioAcousticaR. It is unlikely that you will have to use it directly unless you are writing your own package.
}
\author{
Ed Baker

Maintainer: Ed Baker <edwbaker@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
